+++ src/template.c
@@ -41,23 +41,23 @@ size_t template_replace(char *dest, size_t dest_len, char *src, telemetry_data *
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$bu", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%d", data->temperature_celsius_100 / 100);
+    snprintf(replacement, sizeof(replacement), "%d", (int) data->temperature_celsius_100 / 100);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$te", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%d", data->internal_temperature_celsius_100 / 100);
+    snprintf(replacement, sizeof(replacement), "%d", (int) data->internal_temperature_celsius_100 / 100);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$ti", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%d", data->humidity_percentage_100 / 100);
+    snprintf(replacement, sizeof(replacement), "%d", (int) data->humidity_percentage_100 / 100);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$hu", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%d", data->pressure_mbar_100 / 100);
+    snprintf(replacement, sizeof(replacement), "%d", (int) data->pressure_mbar_100 / 100);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$pr", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%u", data->gps.time_of_week_millis);
+    snprintf(replacement, sizeof(replacement), "%u", (unsigned int) data->gps.time_of_week_millis);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$tow", replacement);
 
@@ -77,15 +77,15 @@ size_t template_replace(char *dest, size_t dest_len, char *src, telemetry_data *
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$sv", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%05d", data->gps.latitude_degrees_1000000 / 10000);
+    snprintf(replacement, sizeof(replacement), "%05d", (int) data->gps.latitude_degrees_1000000 / 10000);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$lat", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%05d", data->gps.longitude_degrees_1000000 / 10000);
+    snprintf(replacement, sizeof(replacement), "%05d", (int) data->gps.longitude_degrees_1000000 / 10000);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$lon", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%d", data->gps.altitude_mm / 1000);
+    snprintf(replacement, sizeof(replacement), "%d", (int) data->gps.altitude_mm / 1000);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$alt", replacement);
 
@@ -93,11 +93,11 @@ size_t template_replace(char *dest, size_t dest_len, char *src, telemetry_data *
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$gs", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%d", data->gps.climb_cm_per_second / 100);
+    snprintf(replacement, sizeof(replacement), "%d", (int) data->gps.climb_cm_per_second / 100);
     strlcpy(temp, dest, dest_len);
     str_replace(dest, dest_len, temp, "$cl", replacement);
 
-    snprintf(replacement, sizeof(replacement), "%03d", data->gps.heading_degrees_100000 / 100000);
+    snprintf(replacement, sizeof(replacement), "%03d", (int) data->gps.heading_degrees_100000 / 100000);
     strlcpy(temp, dest, dest_len);
     size_t len = str_replace(dest, dest_len, temp, "$he", replacement);
